+++ src/main/java/divinerpg/client/renders/entity/projectile/RenderWildwoodLog.java
@@ -35,7 +35,7 @@ public class RenderWildwoodLog extends EntityRenderer<EntityWildwoodLog> {
             Level level = entity.getLevel();
             if (blockstate != level.getBlockState(entity.blockPosition()) && blockstate.getRenderShape() != RenderShape.INVISIBLE) {
                 p_114637_.pushPose();
-                BlockPos blockpos = new BlockPos(entity.getX(), entity.getBoundingBox().maxY, entity.getZ());
+                BlockPos blockpos = new BlockPos((int) entity.getX(), (int) entity.getBoundingBox().maxY, (int) entity.getZ());
                 p_114637_.translate(-0.5D, 0.0D, -0.5D);
                 var model = this.dispatcher.getBlockModel(blockstate);
                 for (var renderType : model.getRenderTypes(blockstate, RandomSource.create(blockstate.getSeed(entity.blockPosition())), net.minecraftforge.client.model.data.ModelData.EMPTY))
